requirements.txt
numpy==1.26.3
PyYAML==6.0.1
torch==2.0.1
tqdm==4.65.2
tensorboard==2.7.0
